hashx: avoid surprising overhead of enum code() method
This is a very simple change that avoids a surprising performance
pitfall: using the code() method on an enum from another crate
caused a non-inlined function call in code where we otherwise expect
a high level of compiler optimization. Replacing code() with a cast
to u8 avoids this function call and allows more intensive optimization
at the call site.

Signed-off-by: Micah Elizabeth Scott <beth@torproject.org>