tor-hsservice: Have the manager be in charge of expiry time
Add the expiry time (as wall clock time) to publish::IptSet.
The manager is in a good position to know this information.

Provide constants that will be used (for now) for actually calculating
the expiry times.

(Right now the part where the expiry time would actually be calculated
doesn't exist, but it will come soon.)