maint/bump_nodep: Allow $CARGO to contain multiple words
This is a more convenient approach to command overriding.  It's what
we do elsewhere.  I need this because with a script like
"maint/bump_nodep", which edits the tree, I need to write
   CARGO='nailing-cargo -E' maint/bump_nodep ....

(The -E option tells my privsep tool that the command is *supposed* to
edit the tree - and that I am intending to review those changes.)