tor-hsservice: ipt_set: Define the shared data structure
Put a last_descriptor_expiry_including_slop into the IptSet.
Define who is responsible for which data.

The code at the sites that interacts with this needs to be bodged, to
make it still compile.  That code is wrong, and the new API for
sharing the state doesn't exist yet.