CI: Remove amd64 tags
We have a new runner, and apparently to use it better we want to
remove these tags.

TBH this seems rather unprincipled.  In theory, shouldn't we tag with
`amd64` all the jobs that we definitely want to run on that arch, and
the others can run on whatever is the default?  Then in the glorious
distant RISCV future everything will run on riscv64 by default?  But
maybe we wouldn't want to switch that as part of the infrastructure
anyway?

Anyway, for discussion see
  https://gitlab.torproject.org/tpo/tpa/team/-/issues/41295#note_2943129