tor-hsservice: Derive PartialEq for Anonimity and DescEncryptionConfig.
This will enable the descriptor publisher to tell whether it needs to
update the descriptor.

When the publisher is notified of an `OnionServiceConfig` change, it
checks whether the `anonimity` or `encrypt_descriptor` fields have
changed. If they have, it rebuilds and republishes the descriptor
using the new values (note: the publisher changes will be implemented in
a future commit).