tor-hsservice: Add an explicit Idle state for the publisher reactor.
The previous code was correct too: after uploading the descriptor, the
`PublishStatus` is not updated, because the reactor is still in a state
where it has enough information to publish descriptors. Note that simply
"being" in the `UploadScheduled` state isn't enough to trigger another
upload, because `publish_status_rx.next()` (from the `select_biased!` of
the main-loop), blocks until something triggers another state transition
(`AwaitingIpts` -> `UploadScheduled`, or even `UploadScheduled` ->
`UploadScheduled`), so while a third state is not strictly necessary, it
does help with readability.