tor-keymgr: Represent unparsed ssh key as strings.
The underlying representation of an `UnparsedOpenSshKey` is now a
`String`. This will make it easier to support storing public keys in the
keystores: in the future, we will use `PublicKey::from_openssh` to parse
public keys, and `PublicKey::from_openssh` expects a string slice
(unlike `PrivateKey::from_openssh`, which takes a `&[u8]`).