Remove direct dependency on generic-array
Instead of tying ourselves to a particular version of the
generic-array crate, we now always use the version re-exported by
our RustCrypto crates.  This lets us avoid the possibility of
version mismatch.

(Originally I had planned to upgrade to generic-array 1.0, but then
I found that we were not actually using it.)