Update changed_crates and Release.md to robustify release process
In particular, this change should help avoid the kind of problem we
ran into as #1054 and fixed with 0c2021c78dc37210, where a crate
changed only in which versions of our crates it depended upon,
and we didn't notice that we had to update its version too.

This commit also fixes a couple of minor issues in the
`changed_crates` script, and gives it a verbose mode.