dirclient: Seal the Requestable trait and hide most of its members.
Since none of these methods were invoked from outside
`tor-dirclient` (except for debugging), and since we have had a fair
amount of churn on what we actually want them to be, it seems like a
good idea to use this trick to hide them.  This will let us make
other changes to the actual behavior of Requestable in the future.