hsserive::req: Yes, leave these functions as async and fallible.
Enough of them have turned out to need to be async so far that I
don't think it's a great idea to commit to making any of them
not-async forever.  Additionally, we've written the code that uses
these, and having them be async didn't seem to cause any major
trouble.

By the same "we may want to make these more complicated underneath"
argument, we should IMO leave them as fallible.