tor-hsservice: Use KeyDenotator::decode instead of manually decoding the denotator.
Note that instead of matching each individual component of a key
denotator (e.g. using a glob pattern like `*_*_*`), we now match the
entire denotator (using the `*` pattern) and let `KeyDenotator::decode`
handle its parsing.

In the long run, this approach should reduce code duplication (as
the decoding logic for each type is centralized in its `KeyDenotator`
implementation) and the need for complex glob patterns.

Closes #1070