circuit: On SendMsgAndInstallHandler, tolerate None handler
Previously it was possible for `handler` to be None only when
`msg` was also None, which would make SendMsgAndInstallHandler into
a no-op.  Now, if `msg` is present but `handler` is absent, we
use the previously installed handler, which I think was our
intention.

Without this patch, `Conversation::send_message` simply won't work.

Fixes #1085.