tor-hsservice: Introduce IptStoreError
If we can't store IPT details, we mustn't publish the corresponding
IPT.  But that's not fatal; maybe we can store later.  So introduce a
new IptStoreError type, and change the return value from some
functions that are going to be able to fail that way.

This introduces a new TODO HSS: we need the publisher to be able to
retry after such a failure.