Do not reject INTRO_ESTABLISHED messages with extensions
The spec says that we should ignore unrecognized extensions.
Per discussion at torspec#249, this is still correct.

Additionally, this commit moves responsibility for checking
INTRO_ESTABLISHED messages into IptMsgHandler::handle_msg, to make
sure that the circuit is torn down as soon as possible on a bad
reply.  (There's nothing to check yet, but there will be once we
start sending extensions that expect a reply.)

Closes #1238.