tor-hsservice: Remove some TODOs about setting the OnionService status.
Currently, the status of the `OnionService` is derived from the statuses
of its subcomponents. Their initial status will be `Shutdown`, so we
don't need to explicitly initialize it.

As for setting the status when `RunnionOnionService::launch` fails, I
don't think that's necessary: `launch()` returns an error on failure, so
the user will know about it (it doesn't need to watch the stream of
status events to find out it failed).

Part of #1083