fslock-guard: Rename os modules
These two modules are key to the implementation of the locking
protocols.

We must define the locking protocol in terms of underlying OS
semantics (since Unix and Windows have different fs concepts and
different concurrency semantics) and therefore, although we are
sharing some code between the implementations, these modules are what
defines the two protocols.