tor-hsservice: key expiry: Do publisher expiry from reactor
In principle there is a small race with doing this in a separate task:
the reactor and the sweeper might process the new netdir at different
times.

For example, if two netdir updates come in quick succession, and the
sweeper is slower, the sweeper might still be running with the
previous TPs as the publisher is creating keys in the new TPs.
Theoretically, the sweeper might delete keys the reactor has just
generated.

This is also considerably simpler.