tor-hsservice: Include the IptManager errors in the onion service status.
The `IptManager` now reports its status as `Recovering` (with the
`latest_error` set to `IptError::FaultyIpt`) if it doesn't have enough
good IPTs to publish. Its status is `Degraded` if it has published some
IPTs, but needs to establish more.

Note: unconditionally setting `latest_error` to `IptError::FaultyIpt` is
not entirely correct. We will need to get the actual errors, if any,
from the `IptEstablisher`.

Part of #1083