arti config watch: Pass a Runtime, not a TorClient
I found this confusing in !2009, where the variable is unused on
Windows: I wondered if this meant our config-reloading only worked on
Unix, since "obviously it would need the TorClient".

But actually the TorClient was only used *in this argument* for its
runtime.  The actual reconfiguration of the TorClient is doen by it
being in one of the ReconfigurableModule's.

Change the type of the argument for clarity.