tor-circmgr: Make the path module crate-private.
We're not currently using any of the path builders outside of
`tor-circmgr`, so let's make them crate-private for now.

Care must be taken if/when we decide to make them public again. For
instance, the `HsPathBuilder` exposes two path building functions, one
that uses vanguards, and one that doesn't. We want to strongly encourage
the use of the vanguards-aware version of the function whenever the
`vanguards` feature is enabled, without breaking any of its existing
non-vanguard uses.

Prompted by https://gitlab.torproject.org/tpo/core/arti/-/merge_requests/2046#note_3010217