tor-rtmock: Introduce MockTimeCore
It bothered me that when I added a `MockCoarseTimeProvider` to the
mock time providers, I had to manually grep to check that there
weren't places the time was updated where I ought to also update the
coarse time.

Prove that the code is right by hiding all the times together in a
struct that prevents un-synchronised updates.

This is part 1, where we move the fields from simple_time::State to
the new struct.