tor-rpcbase: invoker_ent!: remove buggy unuseable arm
This arm puts parens around its argument.  But the other arm insists
on $func:ident, and also would capture any reasonable input.

This was probably something to do with the (func_expr) syntax.  We
don't need to *fix* this macrology, because in a moment the whole
macro can be made to take $func:expr.