proto: Add generic objects for congestion control
This commit adds the congestion window object, a round trip estimator
(RTT) and a state enum. These 3 entities are used by congestion control
in a generic way that is they are passed and used by any algorithm.

At this commit, they are not used hence the allow deadcode attribute for
now in order to minimize the build warnings.

We also introduce the params.rs file containing the parameters, taken
from consensus, used to configure these objects. They will be exposed to
the tor-cirmgr crate to build the CircParameters. More will come.

This also introduces the congestion/ directory that will contain more
code in future commits.

Related #534

Signed-off-by: David Goulet <dgoulet@torproject.org>