proto: New CongestionControl object and fixed window algorithm
Add the top level CongestionControl object that will allow the circuit
reactor to use it in order to decide if a cell can be sent or not. In
order to be used, it is configured with an algorithm that implements the
CongestionControlAlgorithm trait.

The Fixed Window algorithm is also added which essentially implements
the SENDME logic as arti knows it today with a fixed window size.

The SENDME code has been refactor in two different logical steps in
order to accomodate the future Vegas algorithm for which the congestion
window logic and SENDME validity is seperated.

There is now a SENDME validator that takes care of tracking the tags
(authenticated SENDMEs) and validating them upon reception. Then, if
valid, the window management is passed down the congestion control
algorithm, at this commit, FixedWindow object.

Related #534

Signed-off-by: David Goulet <dgoulet@torproject.org>