RPC: Use a slightly less awful workaround in socks.rs
The problem was that Rust won't let us say
```
 type ConnTarget<R> = Arc<dyn ClientConnectionTarget>;
```
because the R parameter wasn't used.

Previously we solved this by using a macro instead of a type
definition, which is ugly.

I had been thinking previously I would need to declare some kind of
additional wrapper type, and had shrunk from the verbosity.  But
@diziet pointed out that I could just use a 2-tuple unconditionally.

It's still not beautiful, but it is less hideous than before.