tor-circmgr: Return an error if HS circ has an invalid length.
Previously, we'd `debug_assert` that the length of the path is valid.
However, `debug_` asserts are compiled out for release builds, which
means that if we have some code path that triggers the assertion
failure, it will go unnoticed unless our tests happen to exercise it.

It's safer to return an internal error, because we definitely don't want
to proceed if the path is too short (see arti#1400 and arti#1409).

Addresses https://gitlab.torproject.org/tpo/core/arti/-/merge_requests/2154#note_3030820