storage: Remove orphaned files from dir_blobs
This patch removes files from dir_blobs if they are not referenced
from the database, or if their filenames are not valid UTF-8.
(If they were not valid UTF-8, we wouldn't have put them in our
database.)

To ensure that there can't be any race conditions, we only do this
when the file is a bit old.