CI: Bail if unexpectedly using a non-amd64 container
cf https://gitlab.torproject.org/tpo/tpa/team/-/issues/41621, it's
possible to unexpectedly run on a container for a different architecture
than the one requested. This can result in subtle and difficult to debug
issues, e.g. when unexpectedly running in the i386 variant of a
container instead of the expected amd64 variant.