tor-hscrypto: Implement FromStr for HsClientDescEncKey.
We will use this impl for parsing the authorized clients of a hidden
service running in "restricted discovery" mode.

NB: "Restricted discovery" is the new terminology for "services that
require client authorization", i.e. services that implement "client
authorization" as described in rend-spec HS-DESC-ENC[0]. Note we haven't
fully transitioned to the new terminology yet (see #1476).

Part of #1292

[0]: https://spec.torproject.org/rend-spec/hsdesc-encrypt.html#HS-DESC-SECOND-LAYER