maint/common/update-shell-includes: Tolerate includes with a `$`
These are not hardcoded, whatever they are, so they're not the
principal thing we're trying to prevent.  Previously, this would trip
on this in arti.git:maint/cargo-publish:

   maint="$(dirname "$0")"
   ...
   # shellcheck source=maint/crates-io-utils.sh
   source "$maint/crates-io-utils.sh"

Ideally we'd teach this script to be able to check and maintain
includes for scripts that aren't in common/, but I think that's a task
for another day.