arti: Elide the -c help in the expected output of the CLI tests.
The help output of the `-c ` option includes some local paths, which can
be quite long on some platforms, spanning over multiple lines. This
causes the CLI tests to fail, because they expect each of the paths from
the `-c` help to fit on a single line.

To fix this, we can use `trycmd`'s `...` to match as many lines as
needed.

Closes #1509