tor-error: implement HasKind on void::Void and Infallible.
`void::Void` represents the type of an object that can't be
constructed.  It's especially useful as the error type of an
infallible function.  `void::Void` already implements
`std::error::Error`, so all we need to do to use it as a "can't
happen" error type in our code is to have it also implement
`HasKind`.

Also implement HasKind for Infallible while we're at it.