tor-proto: removed CountedHashMap
This was written for use in StreamMap, with an eye towards potentially
reusing it elsewhere.

We no longer use it in StreamMap, so it's currently unused.

Probably best to go ahead and remove it now and resurrect it if we
decide we want it later than to maintain dead code (or to make it more
widely available without more concrete use-cases).