Circuit reactor: remove `'hop_streams` loop
As clippy points out otherwise, this no longer actually loops.
Since the iterator already guarantees that the messages it produces
are ready to send with respect to flow control, and we've already
validated that we can send with respect to the circuit window,
the loop, we can always handle the first item from the iterator.

i.e. this changes a `while let` to an `if let`.