tor-memquota: mq_queue: tests: break out check_zero_claimed and fix bug
Each queue has two `Participation`s, not just one.

This didn't matter in the `lifecycle` test, because one of the two
Participations was never touched since it was cloned, and a fresh
clone starts out with a cache of zero.

But we're about to reuse this code in a context where both
Participations end up with a cache.