tor-keys: Automatically implement keymgr trait
The derive ed25519 keypair macro now implements the keymgr trait so the
key wrapper can now be used with a keystore without needing to specify
it in the tor-keymgr crate.

For this to work, a slight change to the KeygenRng trait was needed as
in to expect the CryptoRngCore trait which is what ed25519-dalek
requires.

And also, the removal of the Sealed trait since now it is accepted to
implement these traits outside tor-keymgr.

Fixes #1137

Signed-off-by: David Goulet <dgoulet@torproject.org>