Pass -uc to dpkg-buildpackage when building .debs
By default dpkg-buildpackage signs the output if the d/changelog isn't
UNRELEASED.

If we ever change the d/changelog, we don't want it to fail because it
tries to sign things.

In the future, if and when we want to sign things, that probably won't
be done directly in gitlab CI when the package is built.