tor-chanmgr: add experimental `ChanMgr::handle_incoming`
The channel manager in the future will need to be able to receive
incoming streams. The type of the stream depends on an associated type
within `ChannelFactory`, so this commit exposes this associated type
through several other types, eventually to the `ChanMgr`.

The new methods are behind the experimental "relay" feature flag.