tor_key_forge::traits::ToEncodableKey: add KeyPair associated type
This comment adds a second associated type `KeyPair` to ToEncodableKey.  For a
`ToEncodableKey` which represents a (secret) KeyPair, this type is Self.  For
a `ToEncodableKey` which represents a public key, this is the `ToEncodableKey`
whose `Key` is the pair of which this is the public part.

This is essentially a "type level pointer" from the ToEncodableKey for a
public key to the ToEncodableKey for its secret key.