tor_hsservice use get::<HsIdKey> rather than get::<HsIdKeypair>
There are three places where we query the KeyMgr for an `HsIdKeypair` but all
we really need is the public part.  This commit changes those three callsites
to instead use `get::<HsIdKey>`.

This relies on the previous commit, which makes sure that a request for an
`HsIdKey` will always succeed if the keystore has a `HsIdKeypair` with the
same service nickname.