key-forge: Add helper function and implement traits
Implement the signature::Signer and Ed25519PublicKey trait so the
ed25519 keypair wrapper can be easily used for certificate creation.

This also adds a to_ed25519_id() so we can get a Ed25519Identity which
is an object used around.

Finally, add a prelude module as the list of imports started to grow a
bit out of control.

Part of #1604

Signed-off-by: David Goulet <dgoulet@torproject.org>