tor-socksproto: Factor out return value handling
This centralises the handling of return values, and paves the way for
further deduplication.  (The two return paragraphs were identical.)

There's quite a lot of initial API surface to framework.rs.  Some of
that is going to be hidden, but we're also going to extend this to do
more work.

No functional change.