tor-socksproto: Move `handshake` to be a trait method
This deduplicates some docs and eliminates the two wrapper functiosn
for `run_handshake`, which is now just `handshake`.

We're going to make other API breaks too, and this isn't going to be
the primary API, so we might as well do this.

Proper description of the semver breakage will come at the end when
it's all done.