rpclib: Use prop351 protocol to open streams.
Now that prop351 is what Arti speaks, it's what the rpclib
needs to provide.

Note one change in particular: the `isolation` string
is no longer an optional argument when opening a stream.
(With prop351, there is no longer such a thing as an "absent"
isolation string, and we don't want to imply that there is a
difference between None and "".)