hs-pow: inner doc test using canned data from C tor.
This proliferates the canned hsdesc inner doc testing strategy, adding
another file with data encoded with onion-pow-example running on C
tor. Tests that it parses successfully, and asserts that the pow params
line contents are correct. This is a positive test only.

This strategy seems problematic, but it's better than nothing.

Co-authored-by: Micah Elizabeth Scott <beth@torproject.org>