hs-pow inner desc: broaden pow-params rule, disallow object for v1.
We should not be restricting pow-params to occur only once at the rule
level, and we shouldn't be disallowing object parameters at that level
either. Instead, the v1 scheme itself needs to check for and disallow
objects. Future schemes may allow object parameters.

Test cases for this will be added in a subsequent commit.

Co-authored-by: Micah Elizabeth Scott <beth@torproject.org>